Add BigQuery public dataset
While this resource is centralized (and requires a Google account to access), I believe there's value in knowing about it. Coming from a traditional python dev background, I found BigQuery to be the fastest/simplest/easiest way to analyze transactions. This is my first PR into the website, so any feedback is appreciated.